Update README.md
AutoDynmap automatically fullrenders Dynmap when there are no players online and cancelrenders when a player joins.

This is useful for small private creative Minecraft servers which is what I designed this for. By only rendering when there are no players online, this ensures that the fullrender will capture all of the changes made.

Depending on your settings on Dynmap, This plugin is not useful for servers with a large map and active players 24/7.